+++ tests/test_lanczos.py
@@ -1,14 +1,16 @@
 import numpy as np
-from scipy.linalg import eig   
+from scipy.linalg import eig
 import pytest
 from liteczos.lanczos import get_ground_state, get_green_function
-import liteczos.charlebois as chrlbs
-from test_charlebois import mu_half_filled, U, t
+from liteczos import twosites
+from test_twosites import mu_half, t, U
+
 
 H2x2 = np.array(
     [[-0.09632342, -0.93660398],
      [-0.93660398,  3.73596509]]
 )
+
 
 H16x16 = np.array([
     [-1.3894775 , -1.19471433,  1.99407897,  0.53557267,  0.98367845, -2.62844458,  2.65080982,  0.22244433,  0.14959715, -2.29061152,  0.86372551,  0.82224996,  0.07320181,  1.31539523, -1.95830912, -2.15485695],
@@ -29,13 +31,16 @@ H16x16 = np.array([
     [-2.15485695,  1.58008006, -0.40891196,  0.09134119,  2.81217775,  0.30315493,  1.30594053, -1.59147909,  0.19704743, -0.64377993,  1.61143456, -1.48487055,  1.44579798, -2.06423957, -0.16603588, -2.94753416]
 ])
 
+
 @pytest.fixture(
     name="H",
     params=[H2x2, H16x16],
     ids=["2x2", "16x16"]
 )
 def H(request):
+    
     return request.param
+
 
 @pytest.fixture(
     name="omega",
@@ -45,6 +50,7 @@ def H(request):
 def omega(request):
     return request.param
 
+
 @pytest.fixture(
     name="eta",
     params=[.1, np.linspace(.1, .3, 4)],
@@ -52,6 +58,7 @@ def omega(request):
 )
 def eta(request):
     return request.param
+
 
 @pytest.fixture(
     name="mu",
@@ -76,12 +83,10 @@ def test_ground_state_vector(H):
     assert np.allclose(v0, expected_v0) or np.allclose(v0, -expected_v0)
 
 
-def test_ground_state_charlebois(mu_half_filled, t, U):
-    H = chrlbs.hamiltonian(t, U) - mu_half_filled*chrlbs.number()
-    expected_e0 = chrlbs.ground_state_energy(t, U) - mu_half_filled*2
-
+def test_ground_state_charlebois(mu_half, t, U):
+    H = twosites.hamiltonian(t, U) - mu_half*twosites.number()
+    expected_e0 = twosites.ground_state_energy(t, U) - mu_half*2
     e0, v0 = get_ground_state(H)
-
     assert np.allclose(e0, expected_e0)
     assert np.allclose(H@v0/expected_e0, v0)
